requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiosqlite>=0.19.0
psutil>=5.9.0
